# internal/docs/openapi.yaml
openapi: 3.1.0
info:
    title: ERM Dokter API
    version: 1.0.0
    description: |
        API Elektronik Rekam Medis (ERM) Dokter untuk Rumah Sakit & Klinik berbasis database Khanza.

        ### Fitur Utama:
        - **Authentication**: JWT Bearer Token dengan enkripsi password AES.
        - **Rawat Jalan**: Antrean pasien dokter, pencarian & filter, detail kunjungan pasien.
        - **Pemeriksaan Medis (SOAP & TTV)**: Riwayat pemeriksaan, pencatatan SOAP (Subjective, Objective, Assessment, Plan) baru, update pemeriksaan, dan penghapusan catatan dengan proteksi kepemilikan dokter & batas waktu edit rekam medis (48 jam).
        - **Resep Obat**: Riwayat peresepan obat pasien per kunjungan dan riwayat peresepan seluruh kunjungan berdasarkan No. RM (mendukung filter Ralan, Ranap, Semua, dan rentang tanggal).
        - **Security & Privacy**: Enkripsi identifier URL (`id_kunjungan`, `id_pasien`, `id_pemeriksaan`, `id_obat`) menggunakan AES-GCM URL-safe token untuk mencegah ID enumeration / scraping.
    contact:
        name: Tim Pengembang ERM Dokter
servers:
    - url: http://192.168.30.153:8082
      description: Development Server (LAN / 192.168.30.153)
    - url: http://localhost:8082
      description: Development Server (Localhost)
security:
    - BearerAuth: []
...
components:
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: Masukkan JWT Token yang didapatkan dari endpoint login (/api/v1/auth/login).
    schemas:
...
        PaginationMeta:
            type: object
            properties:
                total_records:
                    type: integer
                    example: 45
                total_pages:
                    type: integer
                    example: 3
                current_page:
                    type: integer
                    example: 1
                per_page:
                    type: integer
                    example: 20
...
        HasilLaboratoriumKunjungan:
            type: object
            properties:
                hasil_pemeriksaan:
                    type: array
                    items:
                        $ref: "#/components/schemas/HasilLaboratorium"
                berkas_digital:
                    type: array
                    items:
                        $ref: "#/components/schemas/BerkasDigital"
paths:
...
    /api/v1/laboratorium/{kategori}/{id_kunjungan}/{status_lanjut}:
        get:
            summary: Riwayat Hasil Laboratorium Kunjungan
            description: Mengambil riwayat hasil pemeriksaan laboratorium pada 1 kunjungan pasien berdasarkan kategori (PK, PA, MB) dan status lanjut (Semua, Ralan, Ranap).
            tags:
                - Laboratorium
            security:
                - BearerAuth: []
            parameters:
                - name: kategori
                  in: path
                  required: true
                  schema:
                      type: string
                      enum: [PK, PA, MB]
                      example: PK
                - name: id_kunjungan
                  in: path
                  required: true
                  schema:
                      type: string
                      example: "kR2mZ9..."
                - name: status_lanjut
                  in: path
                  required: true
                  schema:
                      type: string
                      enum: [Semua, Ralan, Ranap]
                      example: Semua
                - name: tanggal
                  in: query
                  required: false
                  description: Rentang tanggal (YYYY-MM-DD,YYYY-MM-DD)
                  schema:
                      type: string
                      example: "2026-04-01,2026-04-30"
                - name: page
                  in: query
                  schema:
                      type: integer
                      default: 1
                - name: limit
                  in: query
                  schema:
                      type: integer
                      default: 5
            responses:
                "200":
                    description: Berhasil mengambil riwayat hasil laboratorium
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: integer
                                        example: 200
                                    status:
                                        type: string
                                        example: OK
                                    message:
                                        type: string
                                        example: Berhasil mengambil riwayat hasil laboratorium
                                    data:
                                        $ref: "#/components/schemas/HasilLaboratoriumKunjungan"
                                    meta:
                                        $ref: "#/components/schemas/PaginationMeta"
                "400":
                    description: Validasi parameter gagal
    /api/v1/laboratorium/{kategori}/pasien/{id_pasien}/{status_lanjut}:
        get:
            summary: Riwayat Hasil Laboratorium Pasien (Semua Kunjungan by RM)
            description: Mengambil riwayat hasil pemeriksaan laboratorium seluruh kunjungan pasien berdasarkan No RM.
            tags:
                - Laboratorium
            security:
                - BearerAuth: []
            parameters:
                - name: kategori
                  in: path
                  required: true
                  schema:
                      type: string
                      enum: [PK, PA, MB]
                      example: PK
                - name: id_pasien
                  in: path
                  required: true
                  schema:
                      type: string
                      example: "pS9xN1..."
                - name: status_lanjut
                  in: path
                  required: true
                  schema:
                      type: string
                      enum: [Semua, Ralan, Ranap]
                      example: Semua
                - name: tanggal
                  in: query
                  required: false
                  description: Rentang tanggal (YYYY-MM-DD,YYYY-MM-DD)
                  schema:
                      type: string
                      example: "2026-04-01,2026-04-30"
                - name: page
                  in: query
                  schema:
                      type: integer
                      default: 1
                - name: limit
                  in: query
                  schema:
                      type: integer
                      default: 5
            responses:
                "200":
                    description: Berhasil mengambil riwayat hasil laboratorium pasien
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: integer
                                        example: 200
                                    status:
                                        type: string
                                        example: OK
                                    message:
                                        type: string
                                        example: Berhasil mengambil riwayat hasil laboratorium pasien
                                    data:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/HasilLaboratorium"
                                    meta:
                                        $ref: "#/components/schemas/PaginationMeta"
    /api/v1/laboratorium/{kategori}/{id_kunjungan}/{status_lanjut}/{id_hasil}:
        get:
            summary: Detail Hasil Laboratorium
            description: Mengambil detail spesifik hasil pemeriksaan laboratorium beserta rincian parameter atau teks PA dan berkas digital.
            tags:
                - Laboratorium
            security:
                - BearerAuth: []
            parameters:
                - name: kategori
                  in: path
                  required: true
                  schema:
                      type: string
                      enum: [PK, PA, MB]
                      example: PK
                - name: id_kunjungan
                  in: path
                  required: true
                  schema:
                      type: string
                      example: "kR2mZ9..."
                - name: status_lanjut
                  in: path
                  required: true
                  schema:
                      type: string
                      enum: [Semua, Ralan, Ranap]
                      example: Semua
                - name: id_hasil
                  in: path
                  required: true
                  schema:
                      type: string
                      example: "hS8xP1..."
            responses:
                "200":
                    description: Berhasil mengambil detail hasil laboratorium
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: integer
                                        example: 200
                                    status:
                                        type: string
                                        example: OK
                                    message:
                                        type: string
                                        example: Berhasil mengambil detail hasil laboratorium
                                    data:
                                        $ref: "#/components/schemas/HasilLaboratorium"
                "400":
                    description: Parameter tidak valid
                "404":
                    description: Data tidak ditemukan
...
    /api/v1/laboratorium/pk/permintaan/{id_kunjungan}/{status_lanjut}:
        post:
            summary: Kirim Order Permintaan Laboratorium PK Baru
            description: Mengirimkan order permintaan pemeriksaan laboratorium kategori Patologi Klinik (PK) baru oleh dokter. Mendukung transaksi atomik 3 tabel hierarki Khanza, penomoran urut otomatis dengan retry collision, proteksi batas waktu 48 jam rawat jalan, validasi kamar rawat inap aktif, dan proteksi status bayar penjamin BPJS.
            tags:
                - Laboratorium
            security:
                - BearerAuth: []
            parameters:
                - name: id_kunjungan
                  in: path
                  required: true
                  description: Token URL kunjungan terenkripsi
                  schema:
                    type: string
                    example: "w9K2m1..."
                - name: status_lanjut
                  in: path
                  required: true
                  description: Status jenis pelayanan (Ralan / Ranap)
                  schema:
                    type: string
                    enum: [Ralan, Ranap]
                    example: Ralan
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/SimpanPermintaanLabPKRequest"
            responses:
                "201":
                    description: Berhasil mengirim permintaan laboratorium PK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        example: true
                                    message:
                                        type: string
                                        example: Berhasil mengirim permintaan laboratorium PK
                                    data:
                                        $ref: "#/components/schemas/DetailPermintaanLabPK"
                "400":
                    description: Validasi input gagal / waktu tidak valid / batas 48 jam terlewati
                "404":
                    description: Kunjungan tidak ditemukan
        get:
            summary: Daftar Permintaan Laboratorium PK Kunjungan
            description: Mengambil seluruh riwayat permintaan laboratorium PK yang dibuat pada satu kunjungan pasien berdasarkan status pelayanan.
            tags:
                - Laboratorium
            security:
                - BearerAuth: []
            parameters:
                - name: id_kunjungan
                  in: path
                  required: true
                  description: Token URL kunjungan terenkripsi
                  schema:
                    type: string
                - name: status_lanjut
                  in: path
                  required: true
                  description: Filter status pelayanan (Ralan, Ranap, Semua)
                  schema:
                    type: string
                    enum: [Ralan, Ranap, Semua]
                    example: Semua
            responses:
                "200":
                    description: Berhasil mengambil daftar permintaan laboratorium PK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        example: true
                                    message:
                                        type: string
                                    data:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/PermintaanLabPK"
    /api/v1/laboratorium/pk/permintaan/pasien/{id_pasien}/{status_lanjut}:
        get:
            summary: Riwayat Permintaan Laboratorium PK Pasien by No. RM
            description: Mengambil riwayat seluruh permintaan laboratorium PK pasien dari seluruh kunjungan sebelumnya dengan filter status pelayanan, rentang tanggal, dan paginasi.
            tags:
                - Laboratorium
            security:
                - BearerAuth: []
            parameters:
                - name: id_pasien
                  in: path
                  required: true
                  description: Token URL pasien terenkripsi
                  schema:
                    type: string
                - name: status_lanjut
                  in: path
                  required: true
                  description: Filter status pelayanan (Ralan, Ranap, Semua)
                  schema:
                    type: string
                    enum: [Ralan, Ranap, Semua]
                    example: Semua
                - name: tanggal
                  in: query
                  description: Filter rentang tanggal YYYY-MM-DD,YYYY-MM-DD
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    default: 1
                - name: limit
                  in: query
                  schema:
                    type: integer
                    default: 5
            responses:
                "200":
                    description: Berhasil mengambil riwayat permintaan laboratorium PK pasien
    /api/v1/laboratorium/pk/permintaan/{id_kunjungan}/{status_lanjut}/{id_permintaan}:
        get:
            summary: Detail Permintaan Laboratorium PK
            description: Mengambil rincian lengkap permintaan laboratorium PK beserta item tindakan dan sub-template pengujian yang dipilih dokter.
            tags:
                - Laboratorium
            security:
                - BearerAuth: []
            parameters:
                - name: id_kunjungan
                  in: path
                  required: true
                  schema:
                    type: string
                - name: status_lanjut
                  in: path
                  required: true
                  description: Filter status pelayanan (Ralan, Ranap, Semua)
                  schema:
                    type: string
                    enum: [Ralan, Ranap, Semua]
                    example: Semua
                - name: id_permintaan
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: Berhasil mengambil detail permintaan laboratorium PK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        example: true
                                    message:
                                        type: string
                                    data:
                                        $ref: "#/components/schemas/DetailPermintaanLabPK"
                "404":
                    description: Permintaan tidak ditemukan
        put:
            summary: Update / Edit Permintaan Laboratorium PK
            description: |
                Memperbarui data dan daftar pemeriksaan laboratorium PK yang telah dikirimkan oleh dokter dengan nomor permintaan yang tetap.

                ### Aturan Bisnis & Validasi:
                - **Dokter Pembuat**: Hanya dokter yang membuat permintaan (`kd_dokter_perujuk == claims.KodeDokter`) yang berhak memperbarui.
                - **Proteksi Analis Lab**: Permintaan yang sampelnya sudah diambil (`tgl_sampel != 0000-00-00`) atau sudah selesai keluar hasil (`tgl_hasil != 0000-00-00`) **terkunci permanen dan ditolak**.
                - **Proteksi 48 Jam (Rawat Jalan)**: Batas waktu pengeditan maksimal 48 jam dari waktu registrasi pasien.
                - **Validasi Ranap**: Pasien rawat inap tidak boleh sudah checkout/keluar.
                - **Atomisitas Transaksi**: Header permintaan diupdate dan detail pemeriksaan lab (beserta template pengujian) diganti secara atomik dalam 1 transaksi DB.
            tags:
                - Laboratorium
            security:
                - BearerAuth: []
            parameters:
                - name: id_kunjungan
                  in: path
                  required: true
                  schema:
                      type: string
                - name: status_lanjut
                  in: path
                  required: true
                  description: Status pelayanan (Ralan, Ranap, Semua)
                  schema:
                      type: string
                      enum: [Ralan, Ranap, Semua]
                      example: Semua
                - name: id_permintaan
                  in: path
                  required: true
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/SimpanPermintaanLabPKRequest"
            responses:
                "200":
                    description: Berhasil memperbarui permintaan laboratorium PK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        example: true
                                    message:
                                        type: string
                                        example: Berhasil memperbarui permintaan laboratorium PK
                                    data:
                                        $ref: "#/components/schemas/DetailPermintaanLabPK"
                "400":
                    description: Validasi input gagal atau permintaan telah diproses analis lab
                "403":
                    description: Dilarang memperbarui order dokter lain atau melebihi batas waktu 48 jam
                "404":
                    description: Permintaan tidak ditemukan
        delete:
            summary: Batalkan Permintaan Laboratorium PK
            description: Menghapus permintaan laboratorium PK sebelum sampel diambil oleh petugas laboratorium (proteksi tgl_sampel & tgl_hasil belum terisi) dan hanya dapat dibatalkan oleh dokter yang mengirimkan order tersebut.
            tags:
                - Laboratorium
            security:
                - BearerAuth: []
            parameters:
                - name: id_kunjungan
                  in: path
                  required: true
                  schema:
                    type: string
                - name: status_lanjut
                  in: path
                  required: true
                  description: Status pelayanan (Ralan, Ranap, Semua)
                  schema:
                    type: string
                    enum: [Ralan, Ranap, Semua]
                    example: Semua
                - name: id_permintaan
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: Berhasil membatalkan permintaan laboratorium PK
                "400":
                    description: Permintaan sudah diproses laboratorium
                "403":
                    description: Dilarang membatalkan order dokter lain
                "404":
                    description: Permintaan tidak ditemukan
...
    /api/v1/laboratorium/pa/permintaan/pasien/{id_pasien}/{status_lanjut}:
        get:
            summary: Riwayat Permintaan Laboratorium PA Pasien by No. RM
            description: Mengambil riwayat seluruh permintaan laboratorium PA pasien dari seluruh kunjungan sebelumnya dengan filter status pelayanan, rentang tanggal, dan paginasi.
            tags:
                - Laboratorium
            security:
                - BearerAuth: []
            parameters:
                - name: id_pasien
                  in: path
                  required: true
                  description: Token URL pasien terenkripsi
                  schema:
                    type: string
                - name: status_lanjut
                  in: path
                  required: true
                  description: Filter status pelayanan (Ralan, Ranap, Semua)
                  schema:
                    type: string
                    enum: [Ralan, Ranap, Semua]
                    example: Semua
                - name: tanggal
                  in: query
                  description: Filter rentang tanggal YYYY-MM-DD,YYYY-MM-DD
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    default: 1
                - name: limit
                  in: query
                  schema:
                    type: integer
                    default: 5
            responses:
                "200":
                    description: Berhasil mengambil riwayat permintaan laboratorium PA pasien
...
    /api/v1/laboratorium/mb/permintaan/pasien/{id_pasien}/{status_lanjut}:
        get:
            summary: Riwayat Permintaan Laboratorium MB Pasien by No. RM
            description: Mengambil riwayat seluruh permintaan laboratorium MB pasien dari seluruh kunjungan sebelumnya dengan filter status pelayanan, rentang tanggal, dan paginasi.
            tags:
                - Laboratorium
            security:
                - BearerAuth: []
            parameters:
                - name: id_pasien
                  in: path
                  required: true
                  description: Token URL pasien terenkripsi
                  schema:
                    type: string
                - name: status_lanjut
                  in: path
                  required: true
                  description: Filter status pelayanan (Ralan, Ranap, Semua)
                  schema:
                    type: string
                    enum: [Ralan, Ranap, Semua]
                    example: Semua
                - name: tanggal
                  in: query
                  description: Filter rentang tanggal YYYY-MM-DD,YYYY-MM-DD
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    default: 1
                - name: limit
                  in: query
                  schema:
                    type: integer
                    default: 5
            responses:
                "200":
                    description: Berhasil mengambil riwayat permintaan laboratorium MB pasien
...